+++ nes-asp/src/Experimental/Synopses/Sketches/CountMin.cpp
@@ -79,12 +79,12 @@ void CountMin::getApproximateRecord(const uint64_t handlerIndex, Runtime::Execut
         auto h3SeedsCurRow = (h3SeedsMemRef + row * numberOfCols * entrySize).as<Nautilus::MemRef>();
         auto hash = h3HashFunction->calculateWithState(key, h3SeedsCurRow);
         auto colPos = hash % numberOfCols;
-        combineRowsApproximate->combine(sketchArray[(uint64_t) 0][colPosFirstRow], sketchArray[row][colPos]);
+        combineRowsApproximate->combine(sketchArray[0_u64][colPosFirstRow], sketchArray[row][colPos]);
     }
 
     // Writing the key and the approximation to the record
     outputRecord.write(fieldNameKey, key);
-    combineRowsApproximate->lower(sketchArray[(uint64_t) 0][colPosFirstRow], outputRecord);
+    combineRowsApproximate->lower(sketchArray[0_u64][colPosFirstRow], outputRecord);
 }
 
 void CountMin::setup(const uint64_t handlerIndex, Runtime::Execution::ExecutionContext &ctx) {
@@ -96,8 +96,8 @@ void CountMin::setup(const uint64_t handlerIndex, Runtime::Execution::ExecutionC
 
     auto sketchArrayMemRef = Nautilus::FunctionCall("getSketchArrayProxy", getSketchArrayProxy, opHandler);
     auto sketchArray = Nautilus::Interface::Fixed2DArrayRef(sketchArrayMemRef, entrySize, numberOfCols);
-    for (Nautilus::Value<> row = (uint64_t) 0; row < numberOfRows; row = row + 1) {
-        for (Nautilus::Value<> col = (uint64_t) 0; col < numberOfCols; col = col + 1) {
+    for (Nautilus::Value<> row = 0_u64; row < numberOfRows; row = row + 1) {
+        for (Nautilus::Value<> col = 0_u64; col < numberOfCols; col = col + 1) {
             aggregationFunction->reset(sketchArray[row][col]);
         }
     }
