+++ nes-runtime/src/Execution/Operators/ExecutionContext.cpp
@@ -23,8 +23,8 @@ namespace NES::Runtime::Execution {
 
 ExecutionContext::ExecutionContext(const Value<NES::Nautilus::MemRef>& workerContext,
                                    const Value<NES::Nautilus::MemRef>& pipelineContext)
-    : workerContext(workerContext), pipelineContext(pipelineContext), origin((uint64_t) 0), watermarkTs((uint64_t) 0),
-      currentTs((uint64_t) 0), sequenceNumber((uint64_t) 0) {}
+    : workerContext(workerContext), pipelineContext(pipelineContext), origin(0_u64), watermarkTs(0_u64),
+      currentTs(0_u64), sequenceNumber(0_u64) {}
 
 void* allocateBufferProxy(void* workerContextPtr) {
     if (workerContextPtr == nullptr) {
