+++ nes-runtime/tests/UnitTests/Execution/Expressions/ArithmeticalExpressions/DivExpressionTest.cpp
@@ -83,8 +83,8 @@ TEST_F(DivExpressionTest, divUnsignedIntegers) {
         ASSERT_TRUE(resultValue->getTypeIdentifier()->isType<UInt32>());
     }// UInt64
     {
-        auto resultValue = expression.eval(Value<UInt64>((uint64_t) 4), Value<UInt64>((uint64_t) 4));
-        ASSERT_EQ(resultValue, (uint64_t) 1);
+        auto resultValue = expression.eval(Value<UInt64>(4_u64), Value<UInt64>(4_u64));
+        ASSERT_EQ(resultValue, 1_u64);
         ASSERT_TRUE(resultValue->getTypeIdentifier()->isType<UInt64>());
     }
 }
